requirements.txt
pydantic-ai
httpx
asyncio

# A2A SDK dependencies
a2a-sdk
uvicorn

starlette
